+++ src/sources/source_iterable.jl
@@ -25,9 +25,8 @@ function start{T,S}(iter::EnumerableIterable{T,S})
     return start(iter.source)
 end
 
-function next{T,S}(iter::EnumerableIterable{T,S}, state)
-    source_value, source_next_state = next(iter.source, state)
-    return source_value, source_next_state
+@inline function next{T,S}(iter::EnumerableIterable{T,S}, state)
+    return next(iter.source, state)
 end
 
 function done{T,S}(iter::EnumerableIterable{T,S}, state)